requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
dateparser==1.0.0
facebook-scraper==0.2.23
fake-useragent==0.1.11
idna==2.10
lxml==4.6.2
parse==1.19.0
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-html==0.10.0
six==1.15.0
soupsieve==2.2
tornado==6.1
tqdm==4.59.0
tzlocal==2.1
urllib3==1.26.3
w3lib==1.22.0
websockets==8.1
